Introduce an initial command system
Introduce an initial command system, where at the top level you can have
a bunch of commands, which are either "normal commands" or MainCommands.
- "normal commands" are just commands, which can be executed.
- MainCommands are commands, which hold either "normal commands" or
  MainCommands.
This then allows us to build a command tree with sub-commands at
different levels.

All commands are defined in a tree structure in src/commands.
Here you have both types of commmands. An example for a MainCommand
would be src/commands/decensor, whereas an example for a "normal
command" would be src/commands/decensor/auto/bar.

The top level of commands are executed/called in src/app.js.

The MainCommand class, where MainCommands build upon, is defined in
src/MainCommand.js.

This is a variation of the command system outlined here:
https://discordjs.guide/command-handling/

With this new command system, I also changed the command structure
slightly, so that for automatic detection, you have to specify auto like
you do with manual.
However I hope that this command structure can be easily
personalized/configured with aliases in the future.